requirements.txt
numpy>=1.24
scipy>=1.10
librosa>=0.10
soundfile>=0.12
matplotlib>=3.7
seaborn>=0.12
pandas>=1.5
tensorflow-cpu>=2.12
pytest>=7.4
